+++ pkgs/ffigen/lib/src/code_generator/imports.dart
@@ -8,10 +8,16 @@ import 'writer.dart';
 /// A library import which will be written as an import in the generated file.
 class LibraryImport {
   final String name;
-  final String importPath;
+  final String _importPath;
+  final String? _importPathWhenImportedByPackageObjC;
+
   String prefix;
 
-  LibraryImport(this.name, this.importPath) : prefix = name;
+  LibraryImport(this.name, this._importPath,
+      {String? importPathWhenImportedByPackageObjC})
+      : _importPathWhenImportedByPackageObjC =
+            importPathWhenImportedByPackageObjC,
+        prefix = name;
 
   @override
   bool operator ==(other) {
@@ -20,6 +26,13 @@ class LibraryImport {
 
   @override
   int get hashCode => name.hashCode;
+
+  // The import path, which may be different if this library is being imported
+  // into package:objective_c's generated code.
+  String importPath(bool generateForPackageObjectiveC) {
+    if (!generateForPackageObjectiveC) return _importPath;
+    return _importPathWhenImportedByPackageObjC ?? _importPath;
+  }
 }
 
 /// An imported type which will be used in the generated code.
@@ -75,8 +88,9 @@ class SelfImportedType extends Type {
 
 final ffiImport = LibraryImport('ffi', 'dart:ffi');
 final ffiPkgImport = LibraryImport('pkg_ffi', 'package:ffi/ffi.dart');
-final objcPkgImport =
-    LibraryImport('objc', 'package:objective_c/objective_c.dart');
+final objcPkgImport = LibraryImport(
+    'objc', 'package:objective_c/objective_c.dart',
+    importPathWhenImportedByPackageObjC: '../objective_c.dart');
 final self = LibraryImport('self', '');
 
 final voidType = ImportedType(ffiImport, 'Void', 'void');