+++ pkgs/ffigen/test/native_objc_test/is_instance_test.dart
@@ -9,7 +9,9 @@
 import 'dart:ffi';
 import 'dart:io';
 
+import 'package:objective_c/objective_c.dart';
 import 'package:test/test.dart';
+
 import '../test_utils.dart';
 import 'is_instance_bindings.dart';
 import 'util.dart';