+++ pkgs/ffigen/test/native_objc_test/nullable_test.dart
@@ -8,7 +8,9 @@
 import 'dart:ffi';
 import 'dart:io';
 
+import 'package:objective_c/objective_c.dart';
 import 'package:test/test.dart';
+
 import '../test_utils.dart';
 import 'nullable_bindings.dart';
 import 'util.dart';