+++ pkgs/objective_c/lib/src/ns_string.dart
@@ -1,0 +1,9 @@
+// Copyright (c) 2024, the Dart project authors. Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'objective_c_bindings_generated.dart';
+
+extension StringToNSString on String {
+  NSString toNSString() => NSString(this);
+}